Scanner no longer decides underlying type of 'ScannedDevice'
The definition or specific type of 'ScannedDevice' needs to be a protocol. Otherwise we're not allowing API users to define their own, which I think would be best, so. We had to find a way tor this to happen, and my idea was to go the 'some ScannerDevice' way. However, it's not allowed because we can't do that for closures, and then the generic route doesn't work because Scanner is an objective-c route to abide by the CoreBluetooth Delegate stuff. However to my surprise, the 'any' route does work. It requires a specific downcast on the API user's side, who also decides the specific type of the ScannerDevice protocol, so I think it makes sense. And I was even surprised that we were able to switch all the code to work like the Publisher and so on. Seems like we're on the good track.