Return of 'ScannerDevice'
It's no longer a type, it's just a protocol. And we offer CBPeripheral adoption of said protocol, so if someone doesn't define their own type, we allow them to pursue that route. This also means our existing APIs are 'less awkward'. Though perhaps we should reverse the implementations so that the full implementation is the one with the generic.
Note that: 'any ScannerDevice' works, but not 'some ScannerDevice'.