Removed 'openDevice'
It was not working well - sometimes I'd open a Device and then go back again and it wouldn't work. It might still not work, I'll test. But anyway, we wanted the disclosure indicator instead of the Connect Button, so in that sense this fits in better with what we want for the app.