Removed Sample app dependency from iOS-Common-Libraries
Because Common is now a dependency of the BLE Library itself, so we get it through there now. I think originally we did not make the Memfault library dependent on Common because we did not want to push users of Memfault into also depending on Common. But there's so much work that's gone into Common now that it doesn't make sense not to use it. For us, at least.