Refactor: Cleaner API for AppData using KeyPath(s)
No need for multiple functions that do exactly the same thing. Like, exactly. But, it's not as clean as a win because the connection state code does more via the "connectionStateChanged" call. Still, it's an overall improvement. And a bit of a flex, I'll admit. But it was annoying to have so many functions doing the same thing.